add redshift event subscription (wip)
This is a wip commit with some todo question items. I would like to
do a rebase before opening the PR, so that I send in one clean commit
to the pull request. Also, there is another todo item: I should make
the needed changes to go.mod, but no other changes.